# Add minimum roughness project setting to improve VoxelGI/SDFGI performance