The roughness will be clamped for the purposes of VoxelGI and SDFGI rendering. This avoids rendering sharp reflections, which are very demanding when taking significant portions of the screen.

Other sources of reflection such as ReflectionProbes and environment sky aren't affected.

**Testing project:**

## Benchmark (using testing project)

Note that the test scene only uses smooth materials, so the performance gain demonstrated here is greater than it would be in a typical real world scene.

**OS:** Fedora 36
**GPU:** GeForce GTX 1080
**Resolution:** 2560×1440

### No GI

*For reference purposes.*

306 FPS (3.27 mspf)

### VoxelGI only

VoxelGI min. roughness | Performance
0.0 | 170 FPS (5.88 mspf)
0.1 | 171 FPS (5.84 mspf)
0.2 | 181 FPS (5.52 mspf)
0.3 | 186 FPS (5.37 mspf)
0.4 | 195 FPS (5.12 mspf)
0.5 | 198 FPS (5.05 mspf)
0.6 | 200 FPS (5.00 mspf)
0.7 | 202 FPS (4.95 mspf)
0.8 | 204 FPS (4.90 mspf)
0.9 | 206 FPS (4.85 mspf)
1.0 | 208 FPS (4.81 mspf)

### SDFGI only

***Note:** The sudden increase in performance at 0.2 is due to **all** sharp reflections being disabled within the SDFGI shader.*

SDFGI min. roughness | Performance
0.00 | 122 FPS (8.20 mspf)
0.02 | 124 FPS (8.06 mspf)
0.04 | 125 FPS (8.00 mspf)
0.06 | 127 FPS (7.87 mspf)
0.08 | 128 FPS (7.81 mspf)
0.10 | 130 FPS (7.69 mspf)
0.12 | 132 FPS (7.58 mspf)
0.14 | 134 FPS (7.46 mspf)
0.16 | 135 FPS (7.41 mspf)
0.18 | 136 FPS (7.35 mspf)
0.20 | 166 FPS (6.02 mspf)

### VoxelGI + SDFGI

Min. roughnesses | Performance
VoxelGI: 0.0, SDFGI: 0.00 | 96 FPS (10.42 mspf)
VoxelGI: 0.1, SDFGI: 0.02 | 97 FPS (10.31 mspf)
VoxelGI: 0.2, SDFGI: 0.04 | 101 FPS (9.90 mspf)
VoxelGI: 0.3, SDFGI: 0.06 | 105 FPS (9.52 mspf)
VoxelGI: 0.4, SDFGI: 0.08 | 108 FPS (9.26 mspf)
VoxelGI: 0.5, SDFGI: 0.10 | 110 FPS (9.09 mspf)
VoxelGI: 0.6, SDFGI: 0.12 | 111 FPS (9.00 mspf)
VoxelGI: 0.7, SDFGI: 0.14 | 113 FPS (8.85 mspf)
VoxelGI: 0.8, SDFGI: 0.16 | 115 FPS (8.70 mspf)
VoxelGI: 0.9, SDFGI: 0.18 | 117 FPS (8.55 mspf)
VoxelGI: 1.0, SDFGI: 0.20 | 138 FPS (7.25 mspf)