@clayjohn 2022-07-26
I like the idea! Let's reevaluate performance after reduz updates SDFGI and see if this still has the same benefit.