@jcostello 2022-07-27
> I like the idea! Let's reevaluate performance after reduz updates SDFGI and see if this still has the same benefit.

Is this going to happen in the beta phase?