@Calinou 2022-07-27
> Is this going to happen in the beta phase?

Yes :slightly_smiling_face: