@reduz 2022-08-11
This is very hacky because SDFGI switches from sphere tracing to using the probes at a given angle, so at much that should be disabled directly. If you get the value wrong, it will be random when you fall into the sphere tracing range.

I think this should be done different. In fact, for the rewrite I want to make, I want to use a different way to do roughness when TAA is in use that should be a bit faster.